Chore: moved everything from login to authentication so no more file duplication
