+++ src/frontend/app/src/components/Authentication/LoggedOut.vue
