--- src/frontend/app/src/components/login/TwoFactorApp.vue
@@ -1,58 +1,0 @@
-<template>
-
-  <div class="input-group">
-    <input v-bind:class="errorStyling" v-model="code" type="text" class="form-control" placeholder="2FA code" aria-label="2FA code">
-    <button v-bind:class="errorStyling" class="btn btn-outline-light"  @click="submit">Submit</button>
-    <button v-bind:class="errorStyling" class="btn btn-outline-light" @click="logout">Logout</button>
-  </div>
-</template>
-
-<style>
-
-.errorStyling {
-	border-color: red !important;
-}
-
-</style>
-
-<script lang="ts">
-import { useUserStore } from "@/stores/UserStore";
-import { makeApiCallJson } from "@/utils/ApiCall";
-import { logoutUser } from "@/utils/Login";
-import { defineComponent } from "vue";
-
-export default defineComponent({
-  data() {
-    return {
-      code: "",
-	  error: false
-    };
-  },
-  computed: {
-	  errorStyling() {
-		  console.log("error");
-		  if (this.error === true) {
-		  	this.error = false;
-		  	return 'errorStyling';
-		  }
-	  }
-  },
-  methods: {
-    async submit() {
-      const response = await makeApiCallJson("/2fa/authenticate", "POST", {
-        twoFactorCode: this.code,
-      });
-      if (response.ok) {
-        useUserStore().login();
-      }
-	  else {
-		this.code = "";
-		this.error = true;
-	  }
-    },
-    async logout() {
-      await logoutUser(this.$router);
-    },
-  },
-});
-</script>
