+++ src/frontend/app/src/components/navbar/UserProfile.vue
@@ -34,7 +34,7 @@
 import { defineComponent } from "vue";
 import { logoutUser } from "@/utils/Login";
 import { mapState } from "pinia";
-import  LoggedOut  from "@/components/login/LoggedOut.vue";
+import  LoggedOut  from "@/components/Authentication/LoggedOut.vue";
 import makeApiCall from "@/utils/ApiCall";
 
 export default defineComponent({
